Move Nested field type to document.py
